Fixes code according to modern code style. Fixes a few expression bugs.

+++ Scripts/Engines/MLQuests/Items/AsandosSatchel.cs
@@ -1,32 +1,32 @@
 ﻿namespace Server.Items
 {
-	public class AsandosSatchel : Backpack
-	{
-		[Constructible]
-		public AsandosSatchel()
-		{
-			Hue = Utility.RandomBrightHue();
-			DropItem( new SackFlour() );
-			DropItem( new Skillet() );
-		}
+  public class AsandosSatchel : Backpack
+  {
+    [Constructible]
+    public AsandosSatchel()
+    {
+      Hue = Utility.RandomBrightHue();
+      DropItem(new SackFlour());
+      DropItem(new Skillet());
+    }
 
-		public AsandosSatchel( Serial serial )
-			: base( serial )
-		{
-		}
+    public AsandosSatchel(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
