+++ Scripts/Engines/MLQuests/Items/GervisSatchel.cs
@@ -1,32 +1,32 @@
 ﻿namespace Server.Items
 {
-	public class GervisSatchel : Backpack
-	{
-		[Constructible]
-		public GervisSatchel()
-		{
-			Hue = Utility.RandomBrightHue();
-			DropItem( new IronIngot( 10 ) );
-			DropItem( new SmithHammer() );
-		}
+  public class GervisSatchel : Backpack
+  {
+    [Constructible]
+    public GervisSatchel()
+    {
+      Hue = Utility.RandomBrightHue();
+      DropItem(new IronIngot(10));
+      DropItem(new SmithHammer());
+    }
 
-		public GervisSatchel( Serial serial )
-			: base( serial )
-		{
-		}
+    public GervisSatchel(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
