+++ Scripts/Engines/MLQuests/Items/HargroveSatchel.cs
@@ -1,32 +1,32 @@
 ﻿namespace Server.Items
 {
-	public class HargroveSatchel : Backpack
-	{
-		[Constructible]
-		public HargroveSatchel()
-		{
-			Hue = Utility.RandomBrightHue();
-			DropItem( new Gold( 15 ) );
-			DropItem( new Hatchet() );
-		}
+  public class HargroveSatchel : Backpack
+  {
+    [Constructible]
+    public HargroveSatchel()
+    {
+      Hue = Utility.RandomBrightHue();
+      DropItem(new Gold(15));
+      DropItem(new Hatchet());
+    }
 
-		public HargroveSatchel( Serial serial )
-			: base( serial )
-		{
-		}
+    public HargroveSatchel(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
