+++ Scripts/Engines/MLQuests/Items/LowelSatchel.cs
@@ -1,32 +1,32 @@
 ﻿namespace Server.Items
 {
-	public class LowelSatchel : Backpack
-	{
-		[Constructible]
-		public LowelSatchel()
-		{
-			Hue = Utility.RandomBrightHue();
-			DropItem( new Board( 10 ) );
-			DropItem( new DovetailSaw() );
-		}
+  public class LowelSatchel : Backpack
+  {
+    [Constructible]
+    public LowelSatchel()
+    {
+      Hue = Utility.RandomBrightHue();
+      DropItem(new Board(10));
+      DropItem(new DovetailSaw());
+    }
 
-		public LowelSatchel( Serial serial )
-			: base( serial )
-		{
-		}
+    public LowelSatchel(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
