+++ Scripts/Engines/MLQuests/Items/MuggSatchel.cs
@@ -1,32 +1,32 @@
 ﻿namespace Server.Items
 {
-	public class MuggSatchel : Backpack
-	{
-		[Constructible]
-		public MuggSatchel()
-		{
-			Hue = Utility.RandomBrightHue();
-			DropItem( new Pickaxe() );
-			DropItem( new Pickaxe() );
-		}
+  public class MuggSatchel : Backpack
+  {
+    [Constructible]
+    public MuggSatchel()
+    {
+      Hue = Utility.RandomBrightHue();
+      DropItem(new Pickaxe());
+      DropItem(new Pickaxe());
+    }
 
-		public MuggSatchel( Serial serial )
-			: base( serial )
-		{
-		}
+    public MuggSatchel(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
