+++ Scripts/Engines/Quests/The Summoning/Items/ChylothShroud.cs
@@ -1,30 +1,30 @@
 namespace Server.Engines.Quests.Doom
 {
-	public class ChylothShroud : Item
-	{
-		[Constructible]
-		public ChylothShroud() : base( 0x204E )
-		{
-			Hue = 0x846;
-			Layer = Layer.OuterTorso;
-		}
+  public class ChylothShroud : Item
+  {
+    [Constructible]
+    public ChylothShroud() : base(0x204E)
+    {
+      Hue = 0x846;
+      Layer = Layer.OuterTorso;
+    }
 
-		public ChylothShroud( Serial serial ) : base( serial )
-		{
-		}
+    public ChylothShroud(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
