+++ Scripts/Holiday Stuff/Halloween/Treats/Lollipops.cs
@@ -1,38 +1,38 @@
 ﻿namespace Server.Items
 {
-	[TypeAlias( "Server.Items.Lollipop" )]
-	public class Lollipops : CandyCane
-	{
-		[Constructible]
-		public Lollipops()
-			: this( 1 )
-		{
-		}
+  [TypeAlias("Server.Items.Lollipop")]
+  public class Lollipops : CandyCane
+  {
+    [Constructible]
+    public Lollipops()
+      : this(1)
+    {
+    }
 
-		[Constructible]
-		public Lollipops( int amount )
-			: base( 0x468D + Utility.Random( 3 ))
-		{
-			Stackable = true;
-		}
+    [Constructible]
+    public Lollipops(int amount)
+      : base(0x468D + Utility.Random(3))
+    {
+      Stackable = true;
+    }
 
-		public Lollipops( Serial serial )
-			: base( serial )
-		{
-		}
+    public Lollipops(Serial serial)
+      : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( ( int )0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
