+++ Scripts/Items/Addons/StoneFountainAddon.cs
@@ -1,51 +1,51 @@
 namespace Server.Items
 {
-	public class StoneFountainAddon : BaseAddon
-	{
-		[Constructible]
-		public StoneFountainAddon()
-		{
-			int itemID = 0x1731;
+  public class StoneFountainAddon : BaseAddon
+  {
+    [Constructible]
+    public StoneFountainAddon()
+    {
+      int itemID = 0x1731;
 
-			AddComponent( new AddonComponent( itemID++ ), -2, +1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), -1, +1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +0, +1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +1, +1, 0 );
+      AddComponent(new AddonComponent(itemID++), -2, +1, 0);
+      AddComponent(new AddonComponent(itemID++), -1, +1, 0);
+      AddComponent(new AddonComponent(itemID++), +0, +1, 0);
+      AddComponent(new AddonComponent(itemID++), +1, +1, 0);
 
-			AddComponent( new AddonComponent( itemID++ ), +1, +0, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +1, -1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +1, -2, 0 );
+      AddComponent(new AddonComponent(itemID++), +1, +0, 0);
+      AddComponent(new AddonComponent(itemID++), +1, -1, 0);
+      AddComponent(new AddonComponent(itemID++), +1, -2, 0);
 
-			AddComponent( new AddonComponent( itemID++ ), +0, -2, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +0, -1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), +0, +0, 0 );
+      AddComponent(new AddonComponent(itemID++), +0, -2, 0);
+      AddComponent(new AddonComponent(itemID++), +0, -1, 0);
+      AddComponent(new AddonComponent(itemID++), +0, +0, 0);
 
-			AddComponent( new AddonComponent( itemID++ ), -1, +0, 0 );
-			AddComponent( new AddonComponent( itemID++ ), -2, +0, 0 );
+      AddComponent(new AddonComponent(itemID++), -1, +0, 0);
+      AddComponent(new AddonComponent(itemID++), -2, +0, 0);
 
-			AddComponent( new AddonComponent( itemID++ ), -2, -1, 0 );
-			AddComponent( new AddonComponent( itemID++ ), -1, -1, 0 );
+      AddComponent(new AddonComponent(itemID++), -2, -1, 0);
+      AddComponent(new AddonComponent(itemID++), -1, -1, 0);
 
-			AddComponent( new AddonComponent( itemID++ ), -1, -2, 0 );
-			AddComponent( new AddonComponent( ++itemID ), -2, -2, 0 );
-		}
+      AddComponent(new AddonComponent(itemID++), -1, -2, 0);
+      AddComponent(new AddonComponent(++itemID), -2, -2, 0);
+    }
 
-		public StoneFountainAddon( Serial serial ) : base( serial )
-		{
-		}
+    public StoneFountainAddon(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
