+++ Scripts/Items/Body Parts/RightArm.cs
@@ -1,28 +1,28 @@
 namespace Server.Items
 {
-	public class RightArm : Item
-	{
-		[Constructible]
-		public RightArm() : base( 0x1DA2 )
-		{
-		}
+  public class RightArm : Item
+  {
+    [Constructible]
+    public RightArm() : base(0x1DA2)
+    {
+    }
 
-		public RightArm( Serial serial ) : base( serial )
-		{
-		}
+    public RightArm(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
