+++ Scripts/Items/Books/RedBook.cs
@@ -1,43 +1,44 @@
 namespace Server.Items
 {
-	public class RedBook : BaseBook
-	{
-		[Constructible]
-		public RedBook() : base( 0xFF1 )
-		{
-		}
+  public class RedBook : BaseBook
+  {
+    [Constructible]
+    public RedBook() : base(0xFF1)
+    {
+    }
 
-		[Constructible]
-		public RedBook( int pageCount, bool writable ) : base( 0xFF1, pageCount, writable )
-		{
-		}
+    [Constructible]
+    public RedBook(int pageCount, bool writable) : base(0xFF1, pageCount, writable)
+    {
+    }
 
-		[Constructible]
-		public RedBook( string title, string author, int pageCount, bool writable ) : base( 0xFF1, title, author, pageCount, writable )
-		{
-		}
+    [Constructible]
+    public RedBook(string title, string author, int pageCount, bool writable) : base(0xFF1, title, author, pageCount,
+      writable)
+    {
+    }
 
-		// Intended for defined books only
-		public RedBook( bool writable ) : base( 0xFF1, writable )
-		{
-		}
+    // Intended for defined books only
+    public RedBook(bool writable) : base(0xFF1, writable)
+    {
+    }
 
-		public RedBook( Serial serial ) : base( serial )
-		{
-		}
+    public RedBook(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
+      int version = reader.ReadInt();
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int)0 ); // version
-		}
-	}
+      writer.Write(0); // version
+    }
+  }
 }
