+++ Scripts/Items/Champion Artifacts/Decorative/ArtifactLargeVase.cs
@@ -1,28 +1,28 @@
 namespace Server.Items
 {
-	public class ArtifactLargeVase : Item
-	{
-		[Constructible]
-		public ArtifactLargeVase() : base( 0x0B47 )
-		{
-		}
+  public class ArtifactLargeVase : Item
+  {
+    [Constructible]
+    public ArtifactLargeVase() : base(0x0B47)
+    {
+    }
 
-		public ArtifactLargeVase( Serial serial ) : base( serial )
-		{
-		}
+    public ArtifactLargeVase(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
