+++ Scripts/Items/Champion Artifacts/Decorative/SwampTile.cs
@@ -1,28 +1,28 @@
 namespace Server.Items
 {
-	public class SwampTile : Item
-	{
-		[Constructible]
-		public SwampTile() : base( 0x320D )
-		{
-		}
+  public class SwampTile : Item
+  {
+    [Constructible]
+    public SwampTile() : base(0x320D)
+    {
+    }
 
-		public SwampTile( Serial serial ) : base( serial )
-		{
-		}
+    public SwampTile(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
