+++ Scripts/Items/Champion Artifacts/Decorative/WallBlood.cs
@@ -1,29 +1,29 @@
 namespace Server.Items
 {
-	public class WallBlood : Item
-	{
-		[Constructible]
-		public WallBlood()
-			: base(Utility.RandomBool() ? 0x1D95 : 0x1D94)
-		{
-		}
+  public class WallBlood : Item
+  {
+    [Constructible]
+    public WallBlood()
+      : base(Utility.RandomBool() ? 0x1D95 : 0x1D94)
+    {
+    }
 
-		public WallBlood( Serial serial ) : base( serial )
-		{
-		}
+    public WallBlood(Serial serial) : base(serial)
+    {
+    }
 
-		public override void Serialize( GenericWriter writer )
-		{
-			base.Serialize( writer );
+    public override void Serialize(GenericWriter writer)
+    {
+      base.Serialize(writer);
 
-			writer.Write( (int) 0 ); // version
-		}
+      writer.Write(0); // version
+    }
 
-		public override void Deserialize( GenericReader reader )
-		{
-			base.Deserialize( reader );
+    public override void Deserialize(GenericReader reader)
+    {
+      base.Deserialize(reader);
 
-			int version = reader.ReadInt();
-		}
-	}
+      int version = reader.ReadInt();
+    }
+  }
 }
